Search
This project will accept an indexed file and store the
tokens and addresses of the indexed file into memory. It will
then prompt the user to choose which type of search they will
want to use for the program. The two search options are SA
using the and logical operation where it will find the intersect
of the files shared between the inputted tokens and the SO
option will result in the union of the files shared between the
inputted tokens. We used UTHash to construct a hashtable to
store the tokens and their addresses and we then use our
saSearch() and soSearch() methods on the hashtables created
with all the indexed file data to extract the needed information
for the user based on their prompt input.